@@ -2,75 +2,74 @@ using Microsoft.Xna.Framework.Input;
 using TanksRebirth.GameContent;
 using TanksRebirth.Localization;
 
-namespace TanksRebirth
+namespace TanksRebirth;
+
+public class GameConfig
 {
-    public class GameConfig
-    {
-        public float MusicVolume { get; set; } = 0.5f;
-        public float EffectsVolume { get; set; } = 1f;
-        public float AmbientVolume { get; set; } = 1f;
+    public float MusicVolume { get; set; } = 0.5f;
+    public float EffectsVolume { get; set; } = 1f;
+    public float AmbientVolume { get; set; } = 1f;
 
-        #region Graphics Settings
-        public int TankFootprintLimit { get; set; } = 100000;
-        public bool PerPixelLighting { get; set; } = true;
-        public bool Vsync { get; set; } = true;
-        public bool FullScreen { get; set; } = false;
-        public bool MSAA { get; set; } = false;
-        public bool FadeFootprints { get; set; } = false;
+    #region Graphics Settings
+    public int TankFootprintLimit { get; set; } = 100000;
+    public bool PerPixelLighting { get; set; } = true;
+    public bool Vsync { get; set; } = true;
+    public bool FullScreen { get; set; } = false;
+    public bool MSAA { get; set; } = false;
+    public bool FadeFootprints { get; set; } = false;
 
-        public int TargetFPS { get; set; } = 60;
-        #endregion
+    public int TargetFPS { get; set; } = 60;
+    #endregion
 
-        #region Controls Settings
+    #region Controls Settings
 
-        public Keys UpKeybind { get; set; } = Keys.W;
+    public Keys UpKeybind { get; set; } = Keys.W;
 
-        public Keys LeftKeybind { get; set; } = Keys.A;
+    public Keys LeftKeybind { get; set; } = Keys.A;
 
-        public Keys RightKeybind { get; set; } = Keys.D;
+    public Keys RightKeybind { get; set; } = Keys.D;
 
-        public Keys DownKeybind { get; set; } = Keys.S;
+    public Keys DownKeybind { get; set; } = Keys.S;
 
-        public Keys MineKeybind { get; set; } = Keys.Space;
+    public Keys MineKeybind { get; set; } = Keys.Space;
 
-        #endregion
+    #endregion
 
-        #region Res Settings
+    #region Res Settings
 
-        // Defaults to a 4:3 (480p) resolution if not set.
+    // Defaults to a 4:3 (480p) resolution if not set.
 
-        public int ResWidth { get; set; } = 640;
+    public int ResWidth { get; set; } = 640;
 
-        public int ResHeight { get; set; } = 480;
+    public int ResHeight { get; set; } = 480;
 
-        #endregion
+    #endregion
 
-        #region Extra Settings
+    #region Extra Settings
 
-        /// <summary>Used to be casted to a MapTheme to change the... map's theme.</summary>
-        public MapTheme GameTheme { get; set; } = MapTheme.Vanilla;
-        public string MapPack { get; set; } = "Vanilla";
-        public string TankPack { get; set; } = "Vanilla";
-        public string MusicPack { get; set; } = "Vanilla";
+    /// <summary>Used to be casted to a MapTheme to change the... map's theme.</summary>
+    public MapTheme GameTheme { get; set; } = MapTheme.Vanilla;
+    public string MapPack { get; set; } = "Vanilla";
+    public string TankPack { get; set; } = "Vanilla";
+    public string MusicPack { get; set; } = "Vanilla";
 
-        #endregion
+    #endregion
 
-        #region Language
+    #region Language
 
-        public LangCode Language { get; set; } = LangCode.English;
+    public LangCode Language { get; set; } = LangCode.English;
 
-        #endregion
+    #endregion
 
-        // public MultiplayerInfo MultiplayerInfo { get; set; } = default;
-    }
-    public struct MultiplayerInfo
-    {
-        public string Username { get; set; } = "";
+    // public MultiplayerInfo MultiplayerInfo { get; set; } = default;
+}
+public struct MultiplayerInfo
+{
+    public string Username { get; set; } = "";
 
-        public string LastUsedIp { get; set; } = "";
+    public string LastUsedIp { get; set; } = "";
 
-        public string LastUsedPassword { get; set; } = "";
+    public string LastUsedPassword { get; set; } = "";
 
-        public MultiplayerInfo() { }
-    }
+    public MultiplayerInfo() { }
 }
