@@ -4,175 +4,183 @@ using Microsoft.Xna.Framework.Graphics;
 using System;
 using System.Linq;
 using TanksRebirth.GameContent.Systems;
+using TanksRebirth.GameContent.Systems.AI;
 using TanksRebirth.Graphics;
 using TanksRebirth.Internals;
 using TanksRebirth.Internals.Common.Framework.Audio;
 using TanksRebirth.Internals.Common.Utilities;
 
-namespace TanksRebirth.GameContent
-{
-    public class Explosion
-    {
-        // model, blah blah blah
+namespace TanksRebirth.GameContent;
 
-        public delegate void PostUpdateDelegate(Explosion explosion);
-        public static event PostUpdateDelegate OnPostUpdate;
-        public delegate void PostRenderDelegate(Explosion explosion);
-        public static event PostRenderDelegate OnPostRender;
+public class Explosion : IAITankDanger {
+    // model, blah blah blah
 
-        public Tank Source;
+    public delegate void PostUpdateDelegate(Explosion explosion);
+    public static event PostUpdateDelegate? OnPostUpdate;
+    public delegate void PostRenderDelegate(Explosion explosion);
+    public static event PostRenderDelegate? OnPostRender;
 
-        public const int MINE_EXPLOSIONS_MAX = 500;
+    public Tank? Source;
 
-        public static Explosion[] Explosions = new Explosion[MINE_EXPLOSIONS_MAX];
+    public const int MINE_EXPLOSIONS_MAX = 500;
 
-        public Vector2 Position;
+    public static Explosion[] Explosions = new Explosion[MINE_EXPLOSIONS_MAX];
 
-        public bool[] HasHit = new bool[GameHandler.AllTanks.Length];
+    public Vector2 Position { get; set; }
+    public bool IsPlayerSourced { get; set; }
 
-        public Vector3 Position3D => Position.ExpandZ();
+    public bool[] HasHit = new bool[GameHandler.AllTanks.Length];
 
-        public Matrix View;
-        public Matrix Projection;
-        public Matrix World;
+    public Vector3 Position3D => Position.ExpandZ();
 
-        public Model Model;
+    public Matrix View;
+    public Matrix Projection;
+    public Matrix World;
 
-        private static Texture2D _maskingTex;
+    public Model Model;
 
-        public float Scale;
+    private static Texture2D? _maskingTex;
 
-        public float MaxScale;
+    public const float MAGIC_EXPLOSION_NUMBER = 9f;
 
-        public float ExpanseRate = 1f;
-        public float ShrinkRate = 1f;
+    public float Scale;
 
-        public float ShrinkDelay = 40;
+    public float MaxScale;
 
-        private bool _maxAchieved;
+    public float ExpanseRate = 1f;
+    public float ShrinkRate = 1f;
 
-        private int _id;
+    public float ShrinkDelay = 40;
 
-        public float Rotation;
+    private bool _maxAchieved;
 
-        public float RotationSpeed;
+    private int _id;
 
-        public Explosion(Vector2 pos, float scaleMax, Tank owner = null, float rotationSpeed = 1f, float soundPitch = -0.3f)
-        {
-            RotationSpeed = rotationSpeed;
-            Position = pos;
-            MaxScale = scaleMax;
-            Source = owner;
-            _maskingTex = GameResources.GetGameResource<Texture2D>("Assets/textures/misc/tank_smoke_ami");
+    public float Rotation;
 
-            Model = GameResources.GetGameResource<Model>("Assets/mineexplosion");
+    public float RotationSpeed;
 
-            int index = Array.IndexOf(Explosions, null);
+    public Explosion(Vector2 pos, float scaleMax, Tank? owner = null, float rotationSpeed = 1f, float soundPitch = -0.3f) {
+        RotationSpeed = rotationSpeed;
+        Position = pos;
+        MaxScale = scaleMax;
+        Source = owner;
+        _maskingTex = GameResources.GetGameResource<Texture2D>("Assets/textures/misc/tank_smoke_ami");
 
-            var destroysound = "Assets/sounds/tnk_destroy.ogg";
+        AITank.Dangers.Add(this);
+        IsPlayerSourced = owner is null ? false : owner is PlayerTank;
 
-            SoundPlayer.PlaySoundInstance(destroysound, SoundContext.Effect, 1f, 0f, soundPitch, gameplaySound: true);
-            // SoundPlayer.PlaySoundInstance(destroysound, SoundContext.Effect, 0.4f, 0f, -0.2f, gameplaySound: true);
+        Model = GameResources.GetGameResource<Model>("Assets/mineexplosion");
 
-            _id = index;
+        int index = Array.IndexOf(Explosions, null);
 
-            Explosions[index] = this;
-        }
+        var destroysound = "Assets/sounds/tnk_destroy.ogg";
 
-        public void Update() {
-            if (!MapRenderer.ShouldRender)
-                return;
-            if (!_maxAchieved) {
-                if (Scale < MaxScale)
-                    Scale += ExpanseRate * TankGame.DeltaTime;
+        int vertLayers = 5;
+        int horizLayers = 15;
+        for (int i = 0; i < vertLayers; i++) {
+            for (int j = 0; j < horizLayers; j++) {
+                var rot = MathHelper.Tau / horizLayers * i;
 
-                if (Scale > MaxScale)
-                    Scale = MaxScale;
-
-                if (Scale >= MaxScale)
-                    _maxAchieved = true;
             }
-            else if (ShrinkDelay <= 0)
-                Scale -= ShrinkRate * TankGame.DeltaTime;
+        }
 
-            if (!IntermissionSystem.IsAwaitingNewMission) {
-                foreach (var mine in Mine.AllMines) {
-                    if (mine is not null && Vector2.Distance(mine.Position, Position) <= Scale * 9) // magick
-                        mine.Detonate();
-                }
-                foreach (var block in Block.AllBlocks) {
-                    if (block is not null && Vector2.Distance(block.Position, Position) <= Scale * 9 && block.IsDestructible)
-                        block.Destroy();
-                }
-                foreach (var shell in Shell.AllShells) {
-                    if (shell is not null && Vector2.Distance(shell.Position2D, Position) < Scale * 9)
-                        shell.Destroy(Shell.DestructionContext.WithExplosion);
-                }
-                foreach (var tank in GameHandler.AllTanks) {
-                    if (tank is not null && Vector2.Distance(tank.Position, Position) < Scale * 9)
-                        if (!tank.Dead)
-                            if (!HasHit[tank.WorldId])
-                                if (tank.Properties.VulnerableToMines) {
-                                    HasHit[tank.WorldId] = true;
-                                    if (Source is null)
-                                        tank.Damage(new TankHurtContextOther(TankHurtContextOther.HurtContext.FromIngame));
-                                    else if (Source is not null) {
-                                        if (Source is AITank)
-                                            tank.Damage(new TankHurtContextMine(false, this));
-                                        else
-                                            tank.Damage(new TankHurtContextMine(true, this));
-                                    }
-                                }
+        SoundPlayer.PlaySoundInstance(destroysound, SoundContext.Effect, 1f, 0f, soundPitch, gameplaySound: true);
+        // SoundPlayer.PlaySoundInstance(destroysound, SoundContext.Effect, 0.4f, 0f, -0.2f, gameplaySound: true);
+
+        _id = index;
+
+        Explosions[index] = this;
+    }
+
+    public void Update() {
+        if (!MapRenderer.ShouldRenderAll)
+            return;
+        if (!_maxAchieved) {
+            if (Scale < MaxScale)
+                Scale += ExpanseRate * TankGame.DeltaTime;
+
+            if (Scale > MaxScale)
+                Scale = MaxScale;
+
+            if (Scale >= MaxScale)
+                _maxAchieved = true;
+        }
+        else if (ShrinkDelay <= 0)
+            Scale -= ShrinkRate * TankGame.DeltaTime;
+
+        if (!IntermissionSystem.IsAwaitingNewMission) {
+            foreach (var mine in Mine.AllMines) {
+                if (mine is not null && Vector2.Distance(mine.Position, Position) <= Scale * MAGIC_EXPLOSION_NUMBER) // magick
+                    mine.Detonate();
+            }
+            foreach (var block in Block.AllBlocks) {
+                if (block is not null && Vector2.Distance(block.Position, Position) <= Scale * MAGIC_EXPLOSION_NUMBER && block.IsDestructible)
+                    block.Destroy();
+            }
+            foreach (var shell in Shell.AllShells) {
+                if (shell is not null && Vector2.Distance(shell.Position, Position) < Scale * MAGIC_EXPLOSION_NUMBER)
+                    shell.Destroy(Shell.DestructionContext.WithExplosion);
+            }
+            foreach (var tank in GameHandler.AllTanks) {
+                if (tank is null || Vector2.Distance(tank.Position, Position) > Scale * MAGIC_EXPLOSION_NUMBER
+                    || tank.Dead || HasHit[tank.WorldId] || !tank.Properties.VulnerableToMines)
+                    continue;
+                HasHit[tank.WorldId] = true;
+                if (Source is null)
+                    tank.Damage(new TankHurtContextOther(TankHurtContextOther.HurtContext.FromIngame));
+                else if (Source is not null) {
+                    tank.Damage(new TankHurtContextMine(Source is not AITank, this));
                 }
             }
-
-            if (_maxAchieved)
-                ShrinkDelay -= TankGame.DeltaTime;
-
-            if (Scale <= 0)
-                Remove();
-
-            Rotation += RotationSpeed * TankGame.DeltaTime;
-
-            OnPostUpdate?.Invoke(this);
         }
 
-        public void Remove()
-        {
-            Explosions[_id] = null;
-        }
+        if (_maxAchieved)
+            ShrinkDelay -= TankGame.DeltaTime;
 
-        public void Render() {
-            if (!MapRenderer.ShouldRender)
-                return;
-            World = Matrix.CreateScale(Scale) * Matrix.CreateRotationY(Rotation) * Matrix.CreateTranslation(Position3D);
-            View = TankGame.GameView;
-            Projection = TankGame.GameProjection;
+        if (Scale <= 0)
+            Remove();
 
-            foreach (ModelMesh mesh in Model.Meshes) {
-                foreach (BasicEffect effect in mesh.Effects) {
-                    effect.World = World;
-                    effect.View = View;
-                    effect.Projection = Projection;
-                    effect.TextureEnabled = true;
+        Rotation += RotationSpeed * TankGame.DeltaTime;
 
-                    effect.Texture = _maskingTex;
+        OnPostUpdate?.Invoke(this);
+    }
 
-                    effect.SetDefaultGameLighting_IngameEntities();
+    public void Remove() {
+        AITank.Dangers.Remove(this);
+        Explosions[_id] = null;
+    }
 
-                    effect.AmbientLightColor = Color.Orange.ToVector3();
-                    effect.DiffuseColor = Color.Orange.ToVector3();
-                    effect.EmissiveColor = Color.Orange.ToVector3();
-                    effect.FogColor = Color.Orange.ToVector3();
+    public void Render() {
+        if (!MapRenderer.ShouldRenderAll)
+            return;
+        World = Matrix.CreateScale(Scale) * Matrix.CreateRotationY(Rotation) * Matrix.CreateTranslation(Position3D);
+        View = TankGame.GameView;
+        Projection = TankGame.GameProjection;
 
-                    if (ShrinkDelay <= 0)
-                        effect.Alpha -= 0.05f * TankGame.DeltaTime;
-                    else
-                        effect.Alpha = 1f;
-                }
-                mesh.Draw();
+        foreach (ModelMesh mesh in Model.Meshes) {
+            foreach (BasicEffect effect in mesh.Effects) {
+                effect.World = World;
+                effect.View = View;
+                effect.Projection = Projection;
+                effect.TextureEnabled = true;
+
+                effect.Texture = _maskingTex;
+
+                effect.SetDefaultGameLighting_IngameEntities();
+
+                effect.AmbientLightColor = Color.Orange.ToVector3();
+                effect.DiffuseColor = Color.Orange.ToVector3();
+                effect.EmissiveColor = Color.Orange.ToVector3();
+                effect.FogColor = Color.Orange.ToVector3();
+
+                if (ShrinkDelay <= 0)
+                    effect.Alpha -= 0.05f * TankGame.DeltaTime;
+                else
+                    effect.Alpha = 1f;
             }
-            OnPostRender?.Invoke(this);
+            mesh.Draw();
         }
+        OnPostRender?.Invoke(this);
     }
 }
