@@ -12,5 +12,5 @@ public class PingID
     public const int FocusHere = 5;
     public const int GroupHere = 6;
 
-    public static readonly ReflectionDictionary<PingID, int> Collection = new(MemberType.Fields);
+    public static ReflectionDictionary<PingID> Collection { get; internal set; } = new(MemberType.Fields);
 }
