@@ -63,8 +63,7 @@ namespace TanksRebirth.GameContent.Systems
                 }
             }
 
-            foreach (var block in mission.Blocks)
-            {
+            foreach (var block in mission.Blocks) {
                 // TODO: based on this, have a list of defaults per-block type.
                 if (block.Type != BlockID.Hole || block.Type == BlockID.Teleporter)
                     difficulty *= 0.985f;
