@@ -28,7 +28,7 @@ namespace TanksRebirth.Internals.Common.GameUI
             MaxLength = maxLength;
             Id = AllUIElements.IndexOf(AllUIElements.Find(x => x == this));
         }
-        public string GetRealText() => IsEmpty() ? "" : Text;
+        public string GetRealText() => IsEmpty() ? string.Empty : Text.Trim();
         public override void Draw(SpriteBatch spriteBatch) {
             // spriteBatch.DrawString(Font, Text, Position, Color, new Vector2(Scale));
 
