@@ -98,44 +98,44 @@ public static class Easings
         return 1 - InBounce((1 - t) * 2) / 2;
     }
 
-    public static float GetEasingBehavior(EasingType easingType, float easeValue) {
+    public static float GetEasingBehavior(EasingFunction easingType, float easeValue) {
         return easingType switch {
-            EasingType.Linear => Easings.Linear(easeValue),
-            EasingType.InQuad => Easings.InQuad(easeValue),
-            EasingType.OutQuad => Easings.OutQuad(easeValue),
-            EasingType.InOutQuad => Easings.InOutQuad(easeValue),
-            EasingType.InCubic => Easings.InCubic(easeValue),
-            EasingType.OutCubic => Easings.OutCubic(easeValue),
-            EasingType.InOutCubic => Easings.InOutCubic(easeValue),
-            EasingType.InQuart => Easings.InQuart(easeValue),
-            EasingType.OutQuart => Easings.OutQuart(easeValue),
-            EasingType.InOutQuart => Easings.InOutQuart(easeValue),
-            EasingType.InQuint => Easings.InQuint(easeValue),
-            EasingType.OutQuint => Easings.OutQuint(easeValue),
-            EasingType.InOutQuint => Easings.InOutQuint(easeValue),
-            EasingType.InSine => Easings.InSine(easeValue),
-            EasingType.OutSine => Easings.OutSine(easeValue),
-            EasingType.InOutSine => Easings.InOutSine(easeValue),
-            EasingType.InExpo => Easings.InExpo(easeValue),
-            EasingType.OutExpo => Easings.OutExpo(easeValue),
-            EasingType.InOutExpo => Easings.InOutExpo(easeValue),
-            EasingType.InCirc => Easings.InCirc(easeValue),
-            EasingType.OutCirc => Easings.OutCirc(easeValue),
-            EasingType.InOutCirc => Easings.InOutCirc(easeValue),
-            EasingType.InElastic => Easings.InElastic(easeValue),
-            EasingType.OutElastic => Easings.OutElastic(easeValue),
-            EasingType.InOutElastic => Easings.InOutElastic(easeValue),
-            EasingType.InBack => Easings.InBack(easeValue),
-            EasingType.OutBack => Easings.OutBack(easeValue),
-            EasingType.InOutBack => Easings.InOutBack(easeValue),
-            EasingType.InBounce => Easings.InBounce(easeValue),
-            EasingType.OutBounce => Easings.OutBounce(easeValue),
-            EasingType.InOutBounce => Easings.InOutBounce(easeValue),
+            EasingFunction.Linear => Easings.Linear(easeValue),
+            EasingFunction.InQuad => Easings.InQuad(easeValue),
+            EasingFunction.OutQuad => Easings.OutQuad(easeValue),
+            EasingFunction.InOutQuad => Easings.InOutQuad(easeValue),
+            EasingFunction.InCubic => Easings.InCubic(easeValue),
+            EasingFunction.OutCubic => Easings.OutCubic(easeValue),
+            EasingFunction.InOutCubic => Easings.InOutCubic(easeValue),
+            EasingFunction.InQuart => Easings.InQuart(easeValue),
+            EasingFunction.OutQuart => Easings.OutQuart(easeValue),
+            EasingFunction.InOutQuart => Easings.InOutQuart(easeValue),
+            EasingFunction.InQuint => Easings.InQuint(easeValue),
+            EasingFunction.OutQuint => Easings.OutQuint(easeValue),
+            EasingFunction.InOutQuint => Easings.InOutQuint(easeValue),
+            EasingFunction.InSine => Easings.InSine(easeValue),
+            EasingFunction.OutSine => Easings.OutSine(easeValue),
+            EasingFunction.InOutSine => Easings.InOutSine(easeValue),
+            EasingFunction.InExpo => Easings.InExpo(easeValue),
+            EasingFunction.OutExpo => Easings.OutExpo(easeValue),
+            EasingFunction.InOutExpo => Easings.InOutExpo(easeValue),
+            EasingFunction.InCirc => Easings.InCirc(easeValue),
+            EasingFunction.OutCirc => Easings.OutCirc(easeValue),
+            EasingFunction.InOutCirc => Easings.InOutCirc(easeValue),
+            EasingFunction.InElastic => Easings.InElastic(easeValue),
+            EasingFunction.OutElastic => Easings.OutElastic(easeValue),
+            EasingFunction.InOutElastic => Easings.InOutElastic(easeValue),
+            EasingFunction.InBack => Easings.InBack(easeValue),
+            EasingFunction.OutBack => Easings.OutBack(easeValue),
+            EasingFunction.InOutBack => Easings.InOutBack(easeValue),
+            EasingFunction.InBounce => Easings.InBounce(easeValue),
+            EasingFunction.OutBounce => Easings.OutBounce(easeValue),
+            EasingFunction.InOutBounce => Easings.InOutBounce(easeValue),
             _ => throw new Exception()
         };
     }
 }
-public enum EasingType
+public enum EasingFunction
 {
     Linear,
     InQuad, OutQuad, InOutQuad,
